1.) Looked up how to read and write CSV files online. Did a few practice sets before implementing it to my code base. 2.) Created CSVFile class to serve as a means to write and save to a CSV file and use it as a database. 3.) Change toString method of WeeklyScheduleEntry, JournalEntry, and TimeScheduled to use the custom delimiters I made for easy CSV file reading and writing.
Difficulties: Data pulled not showing properly due to bad delimiter used for CSV files. I solved this by adding a unique delimiter that composes of multiple characters which is “-||==||-“. I also made a regular expression version of this pattern since the String.split method takes a regular expression as a parameter.